internal/router: refactor out duplicate code
